requirements.txt
fastapi[standard]
psycopg2
uvicorn[standard]
python-dotenv
pydantic